🏗️ CRITICAL: Fix architecture inconsistencies and security issues
CENTRALIZED SANITIZATION:
- Added sanitize_slug() to agor.utils as canonical implementation
- Removed duplicate sanitize_slug from dev_tools.py
- Updated memory_manager.py to import from utils
- Single source of truth for input sanitization

AGENT DIRECTORY STRUCTURE FIXES:
- Fixed snapshots.py handoff path to use agent directory structure
- Added proper agent_dir definition in create_seamless_handoff()
- Updated file path from .agor/snapshots/ to {agent_dir}snapshots/
- Consistent agent directory usage across all functions

GIT SAFETY IMPROVEMENTS:
- Fixed git add -A to only stage specific removed directories
- Prevents unintentional staging of unrelated files
- Tracks removed directories and stages them explicitly
- Safer git operations during cleanup

ARCHITECTURE CONSISTENCY:
- All functions now use agor/mem/main consistently
- Agent directory structure properly implemented
- Centralized sanitization prevents code divergence
- Secure file operations throughout

TESTING:
✅ Centralized sanitization working
✅ Agent directory structure verified
✅ Memory branch consistency confirmed
✅ Git staging safety implemented
