+++ src/agor/utils.py
@@ -18,6 +18,42 @@ from .exceptions import CompressionError, NetworkError
 from .settings import settings
 
 
+def sanitize_slug(input_string: str) -> str:
+    """
+    Sanitize input string to create safe slugs for file names and branch names.
+
+    Prevents path traversal, injection attacks, and illegal characters.
+
+    Args:
+        input_string: Input string to sanitize
+
+    Returns:
+        Sanitized string containing only safe characters (alphanumerics, dashes, underscores)
+    """
+    import re
+
+    if not input_string:
+        return "unknown"
+
+    # Remove or replace unsafe characters
+    # Keep only alphanumerics, dashes, and underscores
+    sanitized = re.sub(r'[^a-zA-Z0-9\-_]', '_', str(input_string))
+
+    # Remove multiple consecutive underscores/dashes
+    sanitized = re.sub(r'[_\-]+', '_', sanitized)
+
+    # Remove leading/trailing underscores/dashes
+    sanitized = sanitized.strip('_-')
+
+    # Ensure it's not empty and not too long
+    if not sanitized:
+        sanitized = "unknown"
+    elif len(sanitized) > 50:
+        sanitized = sanitized[:50].rstrip('_-')
+
+    return sanitized
+
+
 def move_directory(src_dir: Path, dest_dir: Path):
     dest_dir.mkdir(
         parents=True, exist_ok=True
